Deprecate ddtrace.propagation.utils (#3068)
In preparation for the 1.0 release where we'll remove the utils module.
