+++ ddtrace/propagation/http.py
@@ -12,7 +12,7 @@ from ..internal._tagset import decode_tagset_string
 from ..internal._tagset import encode_tagset_values
 from ..internal.compat import ensure_str
 from ..internal.logger import get_logger
-from .utils import get_wsgi_header
+from ._utils import get_wsgi_header
 
 
 log = get_logger(__name__)
