+++ releasenotes/notes/deprecate_external_ddtrace.propagation.utils-e5106df5b4723310.yaml
@@ -1,0 +1,4 @@
+---
+deprecations:
+  - |
+    ``ddtrace.propagation.utils`` has been deprecated and will be removed in version 1.0.``
